producer/binance/requirements.txt
confluent-kafka==2.3.0
websocket-client==1.7.0
asyncio==3.4.3